Split emoji types
