+++ packages/types/src/discord/auditLogs.ts
@@ -1,7 +1,6 @@
 /** Types for: https://discord.com/developers/docs/resources/audit-log */
 
 import type {
-  DiscordEmoji,
   DiscordGuild,
   DiscordGuildOnboarding,
   DiscordGuildOnboardingPrompt,
@@ -18,6 +17,7 @@ import type {
 } from '../discord.js'
 import type { DiscordAutoModerationRule } from './autoModeration.js'
 import type { DiscordChannel, DiscordOverwrite, DiscordThreadMetadata } from './channels.js'
+import type { DiscordEmoji } from './emojis.js'
 import type { DiscordApplicationCommand, DiscordApplicationCommandPermissions } from './interactions.js'
 
 /** https://discord.com/developers/docs/resources/audit-log#audit-log-object-audit-log-structure */
