+++ packages/types/src/index.ts
@@ -5,6 +5,7 @@ export * from './discord/autoModeration.js'
 export * from './discord/channels.js'
 export * from './discord/gateway.js'
 export * from './discord/interactions.js'
+export * from './discord/emojis.js'
 export * from './discord/reference.js'
 export * from './discord/webhookEvents.js'
 export * from './discordeno.js'
